+++ lib/dropdown_button2.dart
@@ -6,7 +6,6 @@
 */
 
 import 'dart:math' as math;
-import 'dart:ui' show window;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -92,7 +91,7 @@ class _DropdownMenuPainter extends CustomPainter {
 // The widget that is the button wrapping the menu items.
 class _DropdownMenuItemButton<T> extends StatefulWidget {
   const _DropdownMenuItemButton({
-    Key? key,
+    super.key,
     this.padding,
     required this.route,
     required this.buttonRect,
@@ -101,7 +100,7 @@ class _DropdownMenuItemButton<T> extends StatefulWidget {
     required this.enableFeedback,
     this.customItemsIndexes,
     this.customItemsHeight,
-  }) : super(key: key);
+  });
 
   final _DropdownRoute<T> route;
   final EdgeInsets? padding;
@@ -221,7 +220,7 @@ class _DropdownMenuItemButtonState<T>
 
 class _DropdownMenu<T> extends StatefulWidget {
   const _DropdownMenu({
-    Key? key,
+    super.key,
     this.padding,
     required this.route,
     required this.buttonRect,
@@ -236,7 +235,7 @@ class _DropdownMenu<T> extends StatefulWidget {
     required this.offset,
     this.customItemsIndexes,
     this.customItemsHeight,
-  }) : super(key: key);
+  });
 
   final _DropdownRoute<T> route;
   final EdgeInsets? padding;
@@ -351,7 +350,7 @@ class _DropdownMenuState<T> extends State<_DropdownMenu<T>> {
                 child: Scrollbar(
                   radius: widget.scrollbarRadius,
                   thickness: widget.scrollbarThickness,
-                  isAlwaysShown: widget.scrollbarAlwaysShow,
+                  thumbVisibility: widget.scrollbarAlwaysShow,
                   child: ListView(
                     padding: widget.dropdownPadding ?? kMaterialListPadding,
                     shrinkWrap: true,
@@ -676,7 +675,7 @@ class _DropdownRoute<T> extends PopupRoute<_DropdownRouteResult<T>> {
 
 class _DropdownRoutePage<T> extends StatelessWidget {
   const _DropdownRoutePage({
-    Key? key,
+    super.key,
     required this.route,
     required this.constraints,
     this.items,
@@ -699,7 +698,7 @@ class _DropdownRoutePage<T> extends StatelessWidget {
     required this.offset,
     this.customItemsIndexes,
     this.customItemsHeight,
-  }) : super(key: key);
+  });
 
   final _DropdownRoute<T> route;
   final BoxConstraints constraints;
@@ -792,10 +791,10 @@ class _DropdownRoutePage<T> extends StatelessWidget {
 // as closely as possible.
 class _MenuItem<T> extends SingleChildRenderObjectWidget {
   const _MenuItem({
-    Key? key,
+    super.key,
     required this.onLayout,
     required this.item,
-  }) : super(key: key, child: item);
+  }) : super(child: item);
 
   final ValueChanged<Size> onLayout;
   final DropdownMenuItem<T>? item;
@@ -864,7 +863,7 @@ class _DropdownMenuItemContainer extends StatelessWidget {
   }
 }
 
-/// A material design button for selecting from a list of items.
+/// A Material Design button for selecting from a list of items.
 ///
 /// A dropdown button lets the user select from a number of items. The button
 /// shows the currently selected item as well as an arrow that opens a menu for
@@ -925,7 +924,7 @@ class DropdownButton2<T> extends StatefulWidget {
   /// if it is non-null. If [disabledHint] is null, then [hint] will be displayed
   /// if it is non-null.
   DropdownButton2({
-    Key? key,
+    super.key,
     required this.items,
     this.selectedItemBuilder,
     this.value,
@@ -986,11 +985,10 @@ class DropdownButton2<T> extends StatefulWidget {
           'with the same value',
         ),
         _inputDecoration = null,
-        _isEmpty = false,
-        super(key: key);
+        _isEmpty = false;
 
   DropdownButton2._formField({
-    Key? key,
+    super.key,
     required this.items,
     this.selectedItemBuilder,
     this.value,
@@ -1051,8 +1049,7 @@ class DropdownButton2<T> extends StatefulWidget {
           'with the same value',
         ),
         _inputDecoration = inputDecoration,
-        _isEmpty = isEmpty,
-        super(key: key);
+        _isEmpty = isEmpty;
 
   // Parameters added By Me
 
@@ -1316,10 +1313,10 @@ class DropdownButton2<T> extends StatefulWidget {
   final bool _isEmpty;
 
   @override
-  State<DropdownButton2<T>> createState() => _DropdownButton2State<T>();
+  State<DropdownButton2<T>> createState() => DropdownButton2State<T>();
 }
 
-class _DropdownButton2State<T> extends State<DropdownButton2<T>>
+class DropdownButton2State<T> extends State<DropdownButton2<T>>
     with WidgetsBindingObserver {
   int? _selectedIndex;
   _DropdownRoute<T>? _dropdownRoute;
@@ -1356,7 +1353,7 @@ class _DropdownButton2State<T> extends State<DropdownButton2<T>>
 
   @override
   void dispose() {
-    WidgetsBinding.instance!.removeObserver(this);
+    WidgetsBinding.instance.removeObserver(this);
     _removeDropdownRoute();
     focusNode!.removeListener(_handleFocusChanged);
     _internalNode?.dispose();
@@ -1492,6 +1489,10 @@ class _DropdownButton2State<T> extends State<DropdownButton2<T>>
     widget.onMenuStateChange?.call(true);
   }
 
+  // This expose the _handleTap() to Allow opening the button programmatically using GlobalKey.
+  // Also, DropdownButton2State should be public as we need typed access to it through key.
+  void callTap() => _handleTap();
+
   // When isDense is true, reduce the height of this button from _kMenuItemHeight to
   // _kDenseButtonHeight, but don't make it smaller than the text that it contains.
   // Similarly, we don't reduce the height of the button so much that its icon
@@ -1535,7 +1536,7 @@ class _DropdownButton2State<T> extends State<DropdownButton2<T>>
     if (result == null) {
       // If there's no MediaQuery, then use the window aspect to determine
       // orientation.
-      final Size size = window.physicalSize;
+      final Size size = WidgetsBinding.instance.window.physicalSize;
       result = size.width > size.height
           ? Orientation.landscape
           : Orientation.portrait;
@@ -1736,7 +1737,7 @@ class DropdownButtonFormField2<T> extends FormField<T> {
   /// The `items`, `elevation`, `iconSize`, `isDense`, `isExpanded`,
   /// `autofocus`, and `decoration`  parameters must not be null.
   DropdownButtonFormField2({
-    Key? key,
+    super.key,
     required List<DropdownMenuItem<T>>? items,
     DropdownButtonBuilder? selectedItemBuilder,
     T? value,
@@ -1757,8 +1758,8 @@ class DropdownButtonFormField2<T> extends FormField<T> {
     FocusNode? focusNode,
     bool autofocus = false,
     InputDecoration? decoration,
-    FormFieldSetter<T>? onSaved,
-    FormFieldValidator<T>? validator,
+    super.onSaved,
+    super.validator,
     AutovalidateMode? autovalidateMode,
     double? dropdownMaxHeight,
     bool? enableFeedback,
@@ -1799,10 +1800,7 @@ class DropdownButtonFormField2<T> extends FormField<T> {
         ),
         decoration = decoration ?? InputDecoration(focusColor: focusColor),
         super(
-          key: key,
-          onSaved: onSaved,
           initialValue: value,
-          validator: validator,
           autovalidateMode: autovalidateMode ?? AutovalidateMode.disabled,
           builder: (FormFieldState<T> field) {
             final _DropdownButtonFormFieldState<T> state =